+++ Docs/Changes.md
@@ -1,6 +1,8 @@
 # 0.13.0
+  * Change: _Collection object model refactoring._
   * Done: _Arrays, tuples and strings: negative indices_ ([Issue #301](https://github.com/vorov2/dyalect/issues/301), [docs](https://github.com/vorov2/dyalect/wiki/Array#overview), [docs](https://github.com/vorov2/dyalect/wiki/Tuple#overview), [docs](https://github.com/vorov2/dyalect/wiki/String#overview)).
   * Fixed: _Disabling warnings through compiler directive doesn't work_ ([Issue #300](https://github.com/vorov2/dyalect/issues/300)).
+  * Fixed: _Simplified equaility test by `Array.remove`, `Array.indexOf`, `Array.lastIndexof`_ ([Issue #301](https://github.com/vorov2/dyalect/issues/301)).
 
 # 0.12.0
   * Change: _Internal collection (arrays and tuples) object model refactoring._